pyproject.toml
[tool.poetry]
name = "onidata-loan-django"
version = "0.1.0"
description = "Onidata loan challenge django"
authors = ["Victor Augusto <victor.augusto.soliveira@gmail.com>"]
readme = "README.md"
packages = [{ include = "onidata_loan_django" }]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.1"
python-dotenv = "^1.0.0"
dj-database-url = "^2.0.0"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.6"
djangorestframework = "^3.14.0"
markdown = "^3.4.3"
django-filter = "^23.2"

[tool.poetry.group.dev.dependencies]
pytest-django = "*"
flake8 = "*"
black = "*"
isort = "*"
pre-commit = "*"
django-debug-toolbar = "*"
model-bakery = "^1.11.0"
django-extensions = "^3.2.1"
pytest-cov = "^4.0.0"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "onidata.settings"
python_files = ["tests_*.py"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
